modified:   CMakeLists.txt 	modified:   include/calc.h 	modified:   include/fread.h 	modified:   include/layer.h 	modified:   main.c 	modified:   src/calc.c 	modified:   src/fread.c 	modified:   src/layer.c

+++ include/fread.h
@@ -15,7 +15,7 @@ int Freader(MNIST *data);
 
 void FreeMNIST(MNIST *data);
 
-void normalize_imgs(MNIST *data, float *out_images);
+void normalize_imgs(MNIST *data, double *out_images);
 
 
 #endif
